+++ dflib/src/test/java/org/dflib/Window_SelectTest.java
@@ -7,6 +7,13 @@ import static org.dflib.Exp.*;
 
 public class Window_SelectTest {
 
+    static final DataFrame TEST_DF = DataFrame.foldByRow("a", "b").of(
+            1, "x",
+            2, "y",
+            1, "z",
+            0, "a",
+            1, "x");
+
     @Test
     public void empty() {
         DataFrame df = DataFrame.empty("a", "b", "c");
@@ -15,15 +22,8 @@ public class Window_SelectTest {
     }
 
     @Test
-    public void cols_Implicit() {
-        DataFrame df = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y",
-                1, "z",
-                0, "a",
-                1, "x");
-
-        DataFrame r = df.over().select($int("a").sum());
+    public void implicit() {
+        DataFrame r = TEST_DF.over().select($int("a").sum());
         new DataFrameAsserts(r, "sum(a)")
                 .expectHeight(5)
                 .expectRow(0, 5)
@@ -34,40 +34,45 @@ public class Window_SelectTest {
     }
 
     @Test
-    public void cols_ByName() {
-        DataFrame df = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y",
-                1, "z",
-                0, "a",
-                1, "x");
+    public void byName() {
 
-        DataFrame r = df.over()
-                .cols("a", "b")
-                .select(
-                        $int("a").sum(),
-                        $col("b").first()
-                );
+        DataFrame r = TEST_DF.over()
+                .cols("a")
+                .select($int("a").sum());
 
-        new DataFrameAsserts(r, "a", "b")
+        new DataFrameAsserts(r, "a")
                 .expectHeight(5)
-                .expectRow(0, 5, "x")
-                .expectRow(1, 5, "x")
-                .expectRow(2, 5, "x")
-                .expectRow(3, 5, "x")
-                .expectRow(4, 5, "x");
+                .expectRow(0, 5)
+                .expectRow(1, 5)
+                .expectRow(2, 5)
+                .expectRow(3, 5)
+                .expectRow(4, 5);
     }
 
     @Test
-    public void cols_ByPos() {
-        DataFrame df = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y",
-                1, "z",
-                0, "a",
-                1, "x");
+    public void byName_MultiExp() {
 
-        DataFrame r = df.over()
+        DataFrame r = TEST_DF.over()
+                .cols("a", "s", "rn", "cs")
+                .select(
+                        $col("a"), // non-aggregating
+                        $int("a").sum(), // aggregating
+                        rowNum(), // non-aggregating
+                        $int("a").cumSum() // non-aggregating
+                );
+
+        new DataFrameAsserts(r, "a", "s", "rn", "cs")
+                .expectHeight(5)
+                .expectRow(0, 1, 5, 1, 1L)
+                .expectRow(1, 2, 5, 2, 3L)
+                .expectRow(2, 1, 5, 3, 4L)
+                .expectRow(3, 0, 5, 4, 4L)
+                .expectRow(4, 1, 5, 5, 5L);
+    }
+
+    @Test
+    public void byPos() {
+        DataFrame r = TEST_DF.over()
                 .cols(0, 1)
                 .select(
                         $int("a").sum(),
@@ -84,15 +89,8 @@ public class Window_SelectTest {
     }
 
     @Test
-    public void cols_ByPredicate() {
-        DataFrame df = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y",
-                1, "z",
-                0, "a",
-                1, "x");
-
-        DataFrame r = df.over()
+    public void byPredicate() {
+        DataFrame r = TEST_DF.over()
                 .cols("a"::equals)
                 .select($int("a").sum());
 
@@ -106,15 +104,8 @@ public class Window_SelectTest {
     }
 
     @Test
-    public void colsExcept_ByName() {
-        DataFrame df = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y",
-                1, "z",
-                0, "a",
-                1, "x");
-
-        DataFrame r = df.over()
+    public void except_ByName() {
+        DataFrame r = TEST_DF.over()
                 .colsExcept("b")
                 .select($int("a").sum());
 
@@ -128,15 +119,8 @@ public class Window_SelectTest {
     }
 
     @Test
-    public void colsExcept_ByIndex() {
-        DataFrame df = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y",
-                1, "z",
-                0, "a",
-                1, "x");
-
-        DataFrame r = df.over()
+    public void except_ByIndex() {
+        DataFrame r = TEST_DF.over()
                 .colsExcept(1)
                 .select($int("a").sum());
 
@@ -150,15 +134,9 @@ public class Window_SelectTest {
     }
 
     @Test
-    public void colsExcept_ByPredicate() {
-        DataFrame df = DataFrame.foldByRow("a", "b").of(
-                1, "x",
-                2, "y",
-                1, "z",
-                0, "a",
-                1, "x");
+    public void except_ByPredicate() {
 
-        DataFrame r = df.over()
+        DataFrame r = TEST_DF.over()
                 .colsExcept("b"::equals)
                 .select($int("a").sum());